rename src/ManagementRoom.ts -> src/ManagementRoomOutput.ts
